docs(readme): sync with new realities
Discord significantly complicated the process of hosting own bots, to the point that I'm not willing to host one myself. 
As I am not in compliance with their demands, the bot has been purged. 
Hence, I stopped actively supporting the bot.  
This commit updates the readme to reflect these changes.